poke

+++ Skyfall 2.0/Skyfall 2.0/Assets/_Scripts/Bullets/Bullet.cs
@@ -8,8 +8,8 @@ public class Bullet : ScriptableObject {
 
     [Space]
 
-    public int moveSpeed;
+    public float moveSpeed;
     public float lifespan;
 
-    public int maxMoveSpeed = 10;
+    public float maxMoveSpeed = 10f;
 }
